Disallow line appearing more than once
I'm not an expert on Cobertura, but it seems confusing to have stuff
like:

```
<line number="5" hits="1"></line>
<line number="5" hits="0"></line>
```

I decided to treat this situation in a pessimistic way - i.e.: if the
second entry for a line number is lower than the previous one, then we
update the hit count down; but we never update it up. In other words, if
we ever see a 0 for a line, we mark the line uncovered, even if we
previously saw a 1.